hook_event follow-up: replayable events, base hook context, Omit call sites
Follow-up to #37022, addressing the remaining review feedback:

- All events are replayable. broadcastMessage buffers hook_event into the
  SSE replay ring like every other conversation-scoped event — "transient"
  is a rendering contract (clients show it as ephemeral progress), not a
  persistence exception. The replay test now locks the inclusive behavior.

- Base hook context. New assistant/src/hooks/types.ts is the canonical
  home for the chain-hook context shapes: BaseHookContext carries the
  capabilities every chain context shares (logger, broadcast), and
  UserPromptSubmitContext / PostCompactContext / PostToolUseContext /
  StopContext / PreModelCallContext / PostModelCallContext all extend it.
  PluginLogger and HookBroadcast live there too. plugin-api/types.ts
  re-exports everything, so @vellumai/plugin-api imports are unchanged.
  PostCompactContext gains the shared logger (the compact call site
  supplies a request-scoped child logger).

- Omit at call sites. runHook stamps broadcast onto every hook's draft, so
  dispatching call sites construct Omit<XContext, "broadcast"> and never
  supply a placeholder. Applied at the user-prompt-submit site and every
  chain-hook site in agent/loop.ts.
