+++ assistant/src/__tests__/empty-response-hook.test.ts
@@ -94,6 +94,7 @@ function makeCtx(
     stopReason: null,
     decision: "stop",
     logger: noopLogger,
+    broadcast: () => {},
     ...overrides,
   };
 }
@@ -437,6 +438,7 @@ function makeStopCtx(overrides: Partial<StopContext> = {}): StopContext {
     messages: [],
     exitReason: "no_tool_calls",
     logger: noopLogger,
+    broadcast: () => {},
     ...overrides,
   };
 }
