+++ assistant/src/__tests__/exploration-drift-hook.test.ts
@@ -118,6 +118,7 @@ function makeCtx(
     callSite,
     supportsDynamicUi: true,
     logger: noopLogger,
+    broadcast: () => {},
   };
 }
