+++ assistant/src/__tests__/max-tokens-continue-hook.test.ts
@@ -61,6 +61,7 @@ function makeCtx(
     stopReason: "max_tokens",
     decision: "stop",
     logger: noopLogger,
+    broadcast: () => {},
     ...overrides,
   };
 }
