+++ assistant/src/__tests__/task-progress-nudge-hook.test.ts
@@ -95,6 +95,7 @@ function makeCtx(
     callSite: opts.callSite ?? "mainAgent",
     supportsDynamicUi: opts.supportsDynamicUi ?? true,
     logger: noopLogger,
+    broadcast: () => {},
   };
 }
