+++ assistant/src/__tests__/tool-error-hook.test.ts
@@ -84,6 +84,7 @@ function makeCtx(
     callSite: "mainAgent",
     supportsDynamicUi: true,
     logger: noopLogger,
+    broadcast: () => {},
   };
 }
