+++ assistant/src/__tests__/tool-result-truncate-hook.test.ts
@@ -53,6 +53,7 @@ function makeCtx(content: string): PostToolUseContext {
     callSite: "mainAgent",
     supportsDynamicUi: true,
     logger: noopLogger,
+    broadcast: () => {},
   };
 }
